moving ClipContentControl and FrameUpdatingLayoutSection into their own modules
